Albania 2017 Election Results by Polling Station

Data set containing 2017 Albania election results by polling station published by the Central Election Commission and opened by the Coalition of Domestic Observers & Democracy International.

Usage

albania

Format

A data frame with 5362 rows and 45 variables

qarku

district, 12 in total

Q_ID

geocode for district

bashkia

municipality, 61 in total

BAS_ID

geocode for municipality

zaz

election area zone, 90 in total

njesiaAdministrative

village, 373 in total

COM_ID

geocode for village

qvKod

polling station identifier

zgjedhes

number of total registered voters

meshkuj

number of male registered voters

femra

number of female registered voters

totalSeats

number of seats contested by district

vendndodhja

name of polling center containing polling stations

ambienti

type of polling center, 5 in total

totalVoters

number of total registered voters that cast ballots

femVoters

number of female registered voters that cast ballots

maleVoters

number of male registered voters that cast ballots

unusedBallots

number of ballots not used

damagedBallots

number of ballots damaged

ballotsCast

number of total ballots cast

invalidVotes

number of ballots cast that were invalidated

validVotes

number of valid ballots cast

lsi

number of ballots cast for LSI

ps

number of ballots cast for PS

pkd

number of ballots cast for PKD

sfida

number of ballots cast for SFIDA

pr

number of ballots cast for PR

pd

number of ballots cast for PD

pbdksh

number of ballots cast for PBDKSH

adk

number of ballots cast for ADK

psd

number of ballots cast for PSD

ad

number of ballots cast for AD

frd

number of ballots cast for FRD

pds

number of ballots cast for PDS

pdiu

number of ballots cast for PDIU

aak

number of ballots cast for AAK

mega

number of ballots cast for MEGA

pksh

number of ballots cast for PKSH

apd

number of ballots cast for APD

libra

number of ballots cast for LIBRA

psSeats

number of seats won by PS

pdSeats

number of seats won by PD

lsiSeats

number of seats won by LSI

pdiuSeats

number of seats won by PDIU

psdSeats

number of seats won by PSD

Source

https://albaniaelectiondata.herokuapp.com/