Removes duplicate observations within collected data

Removes duplicate observations within collected data

Usage

dedupe(df, col_name)

Arguments

df

object containing data frame of collected data

col_name

variable within data frame by which to filter for duplicate values

Value

Returns table of all data based on unique values within collected data

Examples

aldupe <- rsamp(df=albania, n=390, rep=TRUE)
dedupe(df=aldupe, col_name=qvKod)
#> # A tibble: 378 × 45
#> # Groups:   qvKod [378]
#>    qarku    Q_ID bashkia     BAS_ID zaz   njesiaAdministr… COM_ID qvKod zgjedhes
#>    <fct>   <int> <fct>        <int> <fct> <fct>             <int> <fct>    <int>
#>  1 Durres      3 "Durres"        31 ZAZ … "Durres "          3101 "\"1…      983
#>  2 Tirane     11 "Kamez"        112 ZAZ … "Kamez"           11206 "\"1…      751
#>  3 Tirane     11 "Tirane"       111 ZAZ … "Tirane - Njesi…  11215 "\"1…      777
#>  4 Vlore      12 "Vlore"        128 ZAZ … "Novosele"        12305 "\"4…      720
#>  5 Durres      3 "Durres"        31 ZAZ … "Sukth"            3109 "\"1…      557
#>  6 Tirane     11 "Tirane"       111 ZAZ … "Kashar"          11207 "\"1…      553
#>  7 Fier        5 "Rroskovec"     57 ZAZ … "Strum"            5115 "\"3…      512
#>  8 Fier        5 "Fier "         52 ZAZ … "Fier"             5103 "\"3…      789
#>  9 Shkoder    10 "Shkoder"      105 ZAZ 2 "Ana Malit"       10301 "\"0…      568
#> 10 Elbasan     4 "Gramsh"        43 ZAZ … "Gramsh"           4201 "\"2…      475
#> # … with 368 more rows, and 36 more variables: meshkuj <int>, femra <int>,
#> #   totalSeats <int>, vendndodhja <fct>, ambienti <fct>, totalVoters <int>,
#> #   femVoters <int>, maleVoters <int>, unusedBallots <int>,
#> #   damagedBallots <int>, ballotsCast <int>, invalidVotes <int>,
#> #   validVotes <int>, lsi <int>, ps <int>, pkd <int>, sfida <int>, pr <int>,
#> #   pd <int>, pbdksh <int>, adk <int>, psd <int>, ad <int>, frd <int>,
#> #   pds <int>, pdiu <int>, aak <int>, mega <int>, pksh <int>, apd <int>, …