Identifies missing points between sample and collected data

Identifies missing points between sample and collected data

Usage

rmissing(sampdf, colldf, col_name)

Arguments

sampdf

object containing data frame of sample points

colldf

object containing data frame of collected data

col_name

common variable (i.e. key) in data frames by which to check for missing points

Value

Returns table of sample points missing from collected data

References

Simplified wrapper around dplyr::anti_join()

Examples

alsample <- rsamp(df=albania, 544)
alreceived <- rsamp(df=alsample, 390)
rmissing(sampdf=alsample, colldf=alreceived, col_name=qvKod)
#>          qarku Q_ID        bashkia BAS_ID    zaz
#> 1        Kukes    8        Tropoje     88  ZAZ 9
#> 2         Fier    5        Divjake     51 ZAZ 56
#> 3        Vlore   12         Himare    123 ZAZ 86
#> 4       Tirane   11         Tirane    111 ZAZ 37
#> 5        Vlore   12          Vlore    128 ZAZ 81
#> 6       Tirane   11         Tirane    111 ZAZ 34
#> 7        Berat    1         Kucove     13 ZAZ 63
#> 8        Lezhe    9          Lezhe     92 ZAZ 13
#> 9      Shkoder   10        Shkoder    105  ZAZ 2
#> 10       Kukes    8          Kukes     83 ZAZ 11
#> 11        Fier    5    Mallakaster     55 ZAZ 61
#> 12        Fier    5          Patos     56 ZAZ 59
#> 13       Lezhe    9        Mirdite     94 ZAZ 14
#> 14      Tirane   11         Tirane    111 ZAZ 33
#> 15      Tirane   11           Vore    119 ZAZ 26
#> 16     Elbasan    4        Elbasan     44 ZAZ 48
#> 17       Berat    1          Berat     11 ZAZ 64
#> 18     Shkoder   10 Malesi e Madhe    102  ZAZ 1
#> 19 Gjirokaster    6        Kelcyre     63 ZAZ 75
#> 20       Vlore   12         Himare    123 ZAZ 86
#> 21      Tirane   11         Tirane    111 ZAZ 30
#> 22     Elbasan    4        Elbasan     44 ZAZ 47
#>               njesiaAdministrative COM_ID   qvKod zgjedhes meshkuj femra
#> 1                            Bytyc   8303  "0469"      298     154   144
#> 2                           Terbuf   5216  "2845"      893     473   420
#> 3                           Himare  12303  "4587"      715     354   361
#> 4  Tirane - Njesia Bashkiake Nr. 7  11215  "1920"      695     342   353
#> 5                         Shushice  12310  "4547"      647     324   323
#> 6  Tirane - Njesia Bashkiake Nr. 4  11215  "1804"      963     486   477
#> 7                          Perondi   1203  "3508"      503     259   244
#> 8                         Shenkoll   9208  "0748"      717     366   351
#> 9                        Rrethinat  10310  "0114"      712     373   339
#> 10                           Malzi   8208  "0596"      479     243   236
#> 11                        Aranitas   5301  "3216"      521     284   237
#> 12                           Patos   5110 "30791"      771     404   367
#> 13                           Rubik   9306  "0885"      409     205   204
#> 14 Tirane - Njesia Bashkiake Nr. 3  11215 "17511"      770     386   384
#> 15                            Vore  11217  "1558"      802     437   365
#> 16                         Elbasan   4104 "23232"      989     521   468
#> 17                          Berat    1101  "3312"      312     158   154
#> 18                          Shkrel  10106  "0070"      226     111   115
#> 19                            Suke   6209  "4209"      455     233   222
#> 20                   Hore Vranisht  12313  "4594"      241     128   113
#> 21                      Shengjergj  11214  "1653"      246     127   119
#> 22                   Labinot fushe   4114 "24221"      558     286   272
#>    totalSeats                                                     vendndodhja
#> 1           3                                              Radogoshi, Shkolla
#> 2          16                                           Cerme Sektor, Shkolla
#> 3          12                               Vuno, Ambjent privat "Meri Zisos"
#> 4          34                                            Shkolla "Dhora Leka"
#> 5          12                                                 Risili, Shkolla
#> 6          34                   Rruga "Niko Avrami ", Shkolla " Bajram Curri"
#> 7           7                                 Rreth Tapi, shkolla 9 - Vjecare
#> 8           7                                       Gajush Qendra Shendetsore
#> 9          11                                              Guci e Re, Shkolla
#> 10          3                                   Kalimash, Shkolla 9 - Vjecare
#> 11         16                                 Aranitas 1, Qendra Shendetesore
#> 12         16                         Lagjia '29 Marsi', Shkolla 'Zhani Ciko'
#> 13          7                                        Fang, Shkolla  9-Vjecare
#> 14         34  Rruga "Sitki cico", Shkolla 9-vjecare "Niket Dardani"         
#> 15         34                                                Marikaj, Shkolla
#> 16         14                      Lagjia 5 Maji, Lokal Privat "Sami Shabani"
#> 17          7                          Lagja  "Clirim",  Shkolla "M.A.Virioni
#> 18         11                                     Bzhete Makaj(XHAJ), Shkolla
#> 19          5                                                   Suke, Shkolla
#> 20         12                                             Kallarat, Ambulanca
#> 21         34                                        Fag, shkolla 9 - Vjecare
#> 22         14                                      Mengel, Shkolla 9 -Vjecare
#>    ambienti totalVoters femVoters maleVoters unusedBallots damagedBallots
#> 1    Publik         138        68         70           162              3
#> 2    Publik         364       149        215           542              4
#> 3    Privat         179        73        106           549              1
#> 4    Publik         381       174        207           325              2
#> 5    Publik         259       118        141           400              0
#> 6    Publik         398       183        215           579              5
#> 7    Publik         198        72        126           301              1
#> 8    Publik         294       138        156           437              0
#> 9    Publik         280       132        148           444              0
#> 10   Publik         255       110        145           232              1
#> 11   Publik         268       110        158           263              0
#> 12   Publik         349       160        189           437              0
#> 13   Publik         191        90        101             0              0
#> 14   Publik         386       177        209           399              0
#> 15   Publik         464        10        454           353              1
#> 16   Privat         498       232        266           509              1
#> 17   Publik         158        71         87           155              0
#> 18   Publik          89        38         51           140              1
#> 19   Publik         233       110        123           230              1
#> 20   Publik         135        60         75           109              1
#> 21   Publik         152        58         94            95              3
#> 22   Publik         358       154        204           209              2
#>    ballotsCast invalidVotes validVotes lsi  ps pkd sfida pr  pd pbdksh adk psd
#> 1          138           16        122   4  67   0     0  1  50      0   0   0
#> 2          364            6        358  80 199   0     0  0  57      0   0   1
#> 3          179            2        177  17 144   0     0  0  15      0   0   0
#> 4          381            5        376  43 113   1     1  0 195      0   0   1
#> 5          259            2        257  22 199   0     0  0  26      0   0   1
#> 6          398            8        390  22 207   1     2  2 128      2   0   0
#> 7          198            4        194  58  91   0     0  0  43      1   0   1
#> 8          294            6        288  51  91   1     0  0 134      1   0   0
#> 9          280            7        273  19  95   1     0  1 131      2   0  21
#> 10         255            4        251  26 119   0     0  1 105      0   0   0
#> 11         268            4        264  47 150   0     0  0  53      0   0   0
#> 12         349            6        343  33 200   0     0  1  97      1   1   1
#> 13         191            3        188  16  88   0     0  0  79      0   0   0
#> 14         386           14        372  49 186   0     3  1  85      0   0   0
#> 15         464           10        454  93 194   2     0  2 144      0   0   9
#> 16         498           16        482  46 249   2     0  0  93      0   0   2
#> 17         158            3        155  33  89   0     0  0  25      0   0   0
#> 18          89            1         88   2  50   0     0  0  33      0   0   2
#> 19         233           12        221  74 115   1     0  1  24      0   1   0
#> 20         135            0        135  29  81   0     0  0  25      0   0   0
#> 21         152            4        148  11  57   1     0  0  76      1   0   0
#> 22         358            5        353  30 133   0     1  2 143      0   0   0
#>    ad frd pds pdiu aak mega pksh apd libra psSeats pdSeats lsiSeats pdiuSeats
#> 1   0   0   0    0   0    0    0   0     0       1       2        0         0
#> 2   0   0   1   19   0    0    0   0     1      10       4        2         0
#> 3   0   0   0    0   0    0    0   0     1       8       3        1         0
#> 4   0   1   4    6   0    0    0   2     9      18      11        5         0
#> 5   0   0   0    3   0    0    0   5     1       8       3        1         0
#> 6   0   0   1   11   0    0    1   1    12      18      11        5         0
#> 7   0   0   0    0   0    0    0   0     0       4       1        2         0
#> 8   1   0   0    7   0    0    0   0     2       3       3        1         0
#> 9   0   0   0    3   0    0    0   0     0       4       5        1         0
#> 10  0   0   0    0   0    0    0   0     0       1       2        0         0
#> 11  0   0  11    1   0    0    0   0     2      10       4        2         0
#> 12  0   1   0    8   0    0    0   0     0      10       4        2         0
#> 13  0   0   0    5   0    0    0   0     0       3       3        1         0
#> 14  0   3   0   29   0    0    0   0    16      18      11        5         0
#> 15  0   0   0    3   0    0    0   0     7      18      11        5         0
#> 16  0   0   1   86   0    0    0   0     3       7       3        2         2
#> 17  0   0   1    5   0    0    0   0     2       4       1        2         0
#> 18  0   0   0    0   0    0    0   0     1       4       5        1         0
#> 19  0   0   0    0   0    0    0   0     5       3       1        1         0
#> 20  0   0   0    0   0    0    0   0     0       8       3        1         0
#> 21  0   0   0    1   0    0    0   0     1      18      11        5         0
#> 22  0   4   0   39   0    0    0   0     1       7       3        2         2
#>    psdSeats
#> 1         0
#> 2         0
#> 3         0
#> 4         0
#> 5         0
#> 6         0
#> 7         0
#> 8         0
#> 9         1
#> 10        0
#> 11        0
#> 12        0
#> 13        0
#> 14        0
#> 15        0
#> 16        0
#> 17        0
#> 18        1
#> 19        0
#> 20        0
#> 21        0
#> 22        0
#>  [ reached 'max' / getOption("max.print") -- omitted 132 rows ]