Identifies number of missing points by strata between sample and collected data

Identifies number of missing points by strata between sample and collected data

Usage

smissing(sampdf, colldf, strata, col_name)

Arguments

sampdf

object containing data frame of sample points

colldf

object containing data frame of collected data

strata

variable in both data frames by which to stratify

col_name

common variable (i.e. key) in data frames by which to check for missing points

Value

Returns table of number of sample points by strata missing from collected data

References

Simplified wrapper around dplyr::anti_join()

Examples

alsample <- rsamp(df=albania, 544)
alreceived <- rsamp(df=alsample, 390)
smissing(sampdf=alsample, colldf=alreceived, strata=qarku, col_name=qvKod)
#>          qarku received missing pctreceived pctmissing
#> 1        Berat       23      11    67.64706   32.35294
#> 2        Diber       26       5    83.87097   16.12903
#> 3       Durres       39       8    82.97872   17.02128
#> 4      Elbasan       36      18    66.66667   33.33333
#> 5         Fier       51      17    75.00000   25.00000
#> 6  Gjirokaster       20       6    76.92308   23.07692
#> 7        Korce       32      11    74.41860   25.58140
#> 8        Kukes       10       6    62.50000   37.50000
#> 9        Lezhe       12       4    75.00000   25.00000
#> 10     Shkoder       22      17    56.41026   43.58974
#> 11      Tirane       88      35    71.54472   28.45528
#> 12       Vlore       31      16    65.95745   34.04255