Identifies missing points between sample and collected data

Identifies missing points between sample and collected data

Usage

rmissing(sampdf, colldf, col_name)

Arguments

sampdf

object containing data frame of sample points

colldf

object containing data frame of collected data

col_name

common variable (i.e. key) in data frames by which to check for missing points

Value

Returns table of sample points missing from collected data

References

Simplified wrapper around dplyr::anti_join()

Examples

alsample <- rsamp(df=albania, 544)
alreceived <- rsamp(df=alsample, 390)
rmissing(sampdf=alsample, colldf=alreceived, col_name=qvKod)
#>      qarku Q_ID        bashkia BAS_ID    zaz            njesiaAdministrative
#> 1  Elbasan    4        Elbasan     44 ZAZ 47                     Labinot mal
#> 2     Fier    5      Rroskovec     57 ZAZ 60                           Kuman
#> 3    Vlore   12          Vlore    128 ZAZ 83                           Vlore
#> 4  Elbasan    4         Cerrik     42 ZAZ 46                         Gostime
#> 5     Fier    5          Fier      52 ZAZ 57                          Topoje
#> 6   Durres    3         Durres     31 ZAZ 22                         Durres 
#> 7    Vlore   12          Vlore    128 ZAZ 82                           Vlore
#> 8   Tirane   11         Tirane    111 ZAZ 35 Tirane - Njesia Bashkiake Nr. 5
#> 9    Diber    2          Diber     22 ZAZ 19                       Maqellare
#> 10 Shkoder   10 Malesi e Madhe    102  ZAZ 1                          Koplik
#> 11   Berat    1  Ura Vajgurore     19 ZAZ 62                         Cukalat
#> 12   Lezhe    9        Mirdite     94 ZAZ 14                         Kacinar
#> 13   Vlore   12          Vlore    128 ZAZ 82                           Vlore
#> 14  Tirane   11         Tirane    111 ZAZ 38 Tirane - Njesia Bashkiake Nr. 8
#> 15  Tirane   11         Tirane    111 ZAZ 37 Tirane - Njesia Bashkiake Nr. 7
#> 16   Korce    7       Pogradec     75 ZAZ 67                        Udenisht
#> 17 Shkoder   10        Shkoder    105  ZAZ 2                        Gur i zi
#> 18  Tirane   11         Tirane    111 ZAZ 30                            Dajt
#> 19  Durres    3         Durres     31 ZAZ 23                         Durres 
#> 20    Fier    5          Patos     56 ZAZ 59                           Patos
#> 21 Shkoder   10        Shkoder    105  ZAZ 2                       Rrethinat
#> 22   Korce    7          Korce     73 ZAZ 71                          Mollaj
#>    COM_ID   qvKod zgjedhes meshkuj femra totalSeats
#> 1    4115  "2426"      444     228   216         14
#> 2    5105 "30381"      626     325   301         16
#> 3   12312  "4455"      894     452   442         12
#> 4    4109  "2526"      986     520   466         14
#> 5    5116  "3214"      795     420   375         16
#> 6    3101 "14241"      743     378   365         14
#> 7   12312  "4432"      959     474   485         12
#> 8   11215  "1844"      881     441   440         34
#> 9    2207  "1107"      358     189   169          6
#> 10  10104 "00471"      857     412   445         11
#> 11   1102  "3335"      259     143   116          7
#> 12   9302  "0791"      224     122   102          7
#> 13  12312 "44571"      682     329   353         12
#> 14  11215 "19401"      614     281   333         34
#> 15  11215 "19192"      692     327   365         34
#> 16   7404  "3936"      485     254   231         11
#> 17  10306  "0340"      983     498   485         11
#> 18  11204 "16251"      455     243   212         34
#> 19   3101  "1482"      504     238   266         14
#> 20   5110  "3073"      839     439   400         16
#> 21  10310 "01021"      830     438   392         11
#> 22   7309  "4101"      578     283   295         11
#>                                    vendndodhja ambienti totalVoters femVoters
#> 1                 Labinot Mal, shkolla fillore   Publik         277         0
#> 2                                Luar, Shkolla   Publik         293       149
#> 3                      Zyrat e Nd. Ujesjellesi   Publik         384       170
#> 4                   Gjyrale, Shkolla 9-vjecare   Publik         452       228
#> 5                           Topoje 2,  Shkolle   Publik         358       165
#> 6             Lagjia 14, Shkolla "Isuf Ferra"    Publik         346       151
#> 7                Shkolla "Teli Ndini" - kati I   Publik         345       186
#> 8  Rruga "Kristo Floqi", Shkolla"Besnik Sykja"   Publik         544       255
#> 9                     Dovolan, Shkolla fillore   Publik         215       119
#> 10                          Shkolla 9- vjecare   Publik         316       154
#> 11                           Krotine,  Shkolla   Publik         153        68
#> 12                  Kacinar, Shkolla 9-vjecare   Publik          74        30
#> 13                 Shkolla "28 Nentori" Kati I   Publik         258       124
#> 14                      Shkolla "Skender Caci"   Publik         267       139
#> 15                        Shkolla "Dhora Leka"   Publik         350       171
#> 16                  Piskupat, Shkolla  Fillore   Publik         319       155
#> 17                   Gur i Zi, Shkolla e mesme   Publik         493       241
#> 18                     Priske e Madhe, Shkolla   Publik         266       110
#> 19                         Shkolla "Vasil Ziu"   Publik         176        95
#> 20         Dukas,Lokal privat, Sulejman Rrapaj   Privat         378       161
#> 21                 Dobrac, Qendra Shendetesore   Publik         318       142
#> 22                             Mollaj, Shkolla   Publik         228       109
#>    maleVoters unusedBallots damagedBallots ballotsCast invalidVotes validVotes
#> 1         277             0              0         277           18        259
#> 2         144           344              1         293            7        286
#> 3         214           521              5         384            8        376
#> 4         224           550              3         452           10        442
#> 5         193           452              0         358            5        353
#> 6         195           411              0         346            6        340
#> 7         159           633              0         345            8        337
#> 8         289           352              2         544            5        539
#> 9          96           149              1         215            2        213
#> 10        162           558              1         316            2        314
#> 11         85           108              3         153            3        150
#> 12         44             0              0          74            0         74
#> 13        134           424              0         258            3        255
#> 14        128           359              0         267            6        261
#> 15        179           355              0         350            4        346
#> 16        164           175              0         319            5        314
#> 17        252           505              4         493           11        480
#> 18        156           198              0         266            8        258
#> 19         81           338              0         176            2        174
#> 20        217           476              1         378           10        368
#> 21        176           528              0         318            0        318
#> 22        119           361              0         228            3        225
#>    lsi  ps pkd sfida pr  pd pbdksh adk psd ad frd pds pdiu aak mega pksh apd
#> 1   58 121   0     1  0  48      0   0   0  0   0   0   31   0    0    0   0
#> 2   22 188   0     0  1  66      0   0   0  0   3   1    2   0    0    0   0
#> 3   31 265   0     0  1  67      0   0   0  0   0   0    5   0    0    0   0
#> 4  142 192   0     0  0  92      0   0   0  0   0   0   15   0    0    0   0
#> 5   93 196   0     0  0  50      0   0   0  0   0  12    0   0    0    0   0
#> 6   79 155   0     0  1  84      0   0   0  0   1   0   19   0    0    0   0
#> 7   26 218   0     1  0  70      0   0   0  0   2   0   16   0    0    0   1
#> 8   94 278   0     4  0 121      0   1   1  0   1   0   24   0    0    0   0
#> 9   16 102   0     0  0  56      0   1   0  0   1   0   37   0    0    0   0
#> 10  20 171   0     0  0  79      0   0  35  0   0   0    4   0    0    0   0
#> 11  15  86   0     0  0  47      0   0   0  0   2   0    0   0    0    0   0
#> 12  14  38   0     0  0  18      0   0   0  0   0   0    4   0    0    0   0
#> 13   6 169   0     0  0  67      0   0   0  0   3   0    5   0    0    1   1
#> 14  24 100   0     3  0 118      0   0   0  0   0   0    5   0    0    0   3
#> 15  40 191   0     2  0  98      0   0   1  0   1   0    3   1    0    0   1
#> 16  51 173   0     0  0  84      0   0   1  0   0   1    0   0    0    0   0
#> 17 138  52   1     0  0  99      1   4 156  1   0   0   24   0    0    0   0
#> 18  21 126   2     0  2  78      0   0   0  0   0   0   13   0    0    0   1
#> 19  25  82   0     1  0  40      0   0   0  0   6   0   19   0    0    0   0
#> 20  71 218   2     0  0  64      0   8   0  0   1   1    1   0    0    2   0
#> 21  14 185   0     0  1  82      0   0  34  0   0   0    0   0    0    0   0
#> 22  54 109   0     0  0  60      0   0   0  0   0   0    0   0    0    0   0
#>    libra psSeats pdSeats lsiSeats pdiuSeats psdSeats
#> 1      0       7       3        2         2        0
#> 2      3      10       4        2         0        0
#> 3      7       8       3        1         0        0
#> 4      1       7       3        2         2        0
#> 5      2      10       4        2         0        0
#> 6      1       8       4        2         0        0
#> 7      3       8       3        1         0        0
#> 8     15      18      11        5         0        0
#> 9      0       2       2        1         1        0
#> 10     5       4       5        1         0        1
#> 11     0       4       1        2         0        0
#> 12     0       3       3        1         0        0
#> 13     3       8       3        1         0        0
#> 14     8      18      11        5         0        0
#> 15     8      18      11        5         0        0
#> 16     4       6       4        1         0        0
#> 17     4       4       5        1         0        1
#> 18    15      18      11        5         0        0
#> 19     1       8       4        2         0        0
#> 20     0      10       4        2         0        0
#> 21     2       4       5        1         0        1
#> 22     2       6       4        1         0        0
#>  [ reached 'max' / getOption("max.print") -- omitted 132 rows ]