Identifies number of missing points by strata between sample and collected data

Identifies number of missing points by strata between sample and collected data

Usage

smissing(sampdf, colldf, strata, col_name)

Arguments

sampdf

object containing data frame of sample points

colldf

object containing data frame of collected data

strata

variable in both data frames by which to stratify

col_name

common variable (i.e. key) in data frames by which to check for missing points

Value

Returns table of number of sample points by strata missing from collected data

References

Simplified wrapper around dplyr::anti_join()

Examples

alsample <- rsamp(df=albania, 544)
alreceived <- rsamp(df=alsample, 390)
smissing(sampdf=alsample, colldf=alreceived, strata=qarku, col_name=qvKod)
#>          qarku received missing pctreceived pctmissing
#> 1        Berat       22       6    78.57143   21.42857
#> 2        Diber       19       6    76.00000   24.00000
#> 3       Durres       34      13    72.34043   27.65957
#> 4      Elbasan       39      15    72.22222   27.77778
#> 5         Fier       46      18    71.87500   28.12500
#> 6  Gjirokaster       17       4    80.95238   19.04762
#> 7        Korce       35      16    68.62745   31.37255
#> 8        Kukes        8      10    44.44444   55.55556
#> 9        Lezhe       15       6    71.42857   28.57143
#> 10     Shkoder       32      13    71.11111   28.88889
#> 11      Tirane       95      31    75.39683   24.60317
#> 12       Vlore       28      16    63.63636   36.36364